refactor: 공통 필드 BaseEntity 상속
