+++ src/main/java/com/sprint/mission/discodeit/entity/BaseEntity.java
@@ -1,0 +1,33 @@
+package com.sprint.mission.discodeit.entity;
+
+import com.sprint.mission.discodeit.global.UtilMethod;
+
+import java.util.UUID;
+
+public abstract class BaseEntity {
+    private final UUID id;
+    private final Long createdAt;
+    private Long updatedAt;
+
+    public BaseEntity() {
+        this.id = UUID.randomUUID();
+        this.createdAt = UtilMethod.getCurrentTime();
+        this.updatedAt = 0L;
+    }
+
+    public UUID getId() {
+        return id;
+    }
+
+    public Long getCreatedAt() {
+        return createdAt;
+    }
+
+    public Long getUpdatedAt() {
+        return updatedAt;
+    }
+
+    public void updateUpdatedAt(Long updatedAt) {
+        this.updatedAt = updatedAt;
+    }
+}
