+++ src/main/java/com/sprint/mission/discodeit/entity/User.java
@@ -5,33 +5,25 @@ import com.sprint.mission.discodeit.global.UtilMethod;
 import java.util.List;
 import java.util.UUID;
 
-public class User {
-    private final UUID id;
+public class User extends BaseEntity {
     private String name;
     private String nickname;
     private String email;
     private String password;
     private String profileImageUrl;
     private boolean active;
-    private final Long createdAt;
-    private Long updatedAt;
 
     public User(String name, String nickname, String email, String password,
                 String profileImageUrl, boolean active) {
-        this.id = UUID.randomUUID();
+        super();
         this.name = name;
         this.nickname = nickname;
         this.email = email;
         this.password = password;
         this.profileImageUrl = profileImageUrl;
         this.active = active;
-        this.createdAt = UtilMethod.getCurrentTime();
-        this.updatedAt = 0L;
     }
 
-    public UUID getId() {
-        return id;
-    }
 
     public String getName() {
         return name;
@@ -58,14 +50,6 @@ public class User {
     }
 
 
-    public Long getCreatedAt() {
-        return createdAt;
-    }
-
-    public Long getUpdatedAt() {
-        return updatedAt;
-    }
-
     public void updateName(String name) {
         this.name = name;
     }
@@ -90,7 +74,4 @@ public class User {
         this.active = !this.active;
     }
 
-    public void updateUpdatedAt(Long updatedAt) {
-        this.updatedAt = updatedAt;
-    }
 }
